+++ test/unit/WelcomeTest.java
@@ -1,4 +1,5 @@
 package unit;
+
 import static org.fest.assertions.Assertions.assertThat;
 import static play.test.Helpers.contentAsString;
 import static play.test.Helpers.contentType;
